requirements.txt
typer[all]
rich
gitpython
google-generativeai
pytest
pytest-cov
pytest-mock
pyyaml
streamlit
plotly
pandas
black
isort
flake8
mypy
types-PyYAML
